requirements.txt
# LLM + orchestration
langchain>=0.1.17
langgraph>=0.0.40
langchain_core>=0.1.52

# Groq API
langchain-groq>=0.1.3

# Data models & validation
pydantic>=2.0

# Environment variable management
python-dotenv>=1.0.1
async-timeout>=4.0

# Math utilities (used later for equilibria)
numpy

# Streamlit visualization
streamlit>=1.35.0